[DX] Prefer overmind to foreman
If available, use overmind. Fallback to foreman otherwise (windows,
preference etc).

Overmind is more stable, provides various featues as it's based on tmux,
makes it trivial to stop/restart to/attach to processes (that you
stopped with binding.irb, for example), so over all I feel like it's a
better DX.

Foreman is fine too (it's part of Omakase Rails now), so you can't
really go wrong using it either